refactor and create user service function
