+++ docker-compose.yaml
@@ -7,5 +7,5 @@ services:
       POSTGRES_PASSWORD: shamiri
       POSTGRES_DB: rafi_db
     ports:
-      - 5432:5432
+      - 5433:5432
     command: postgres -c log_destination=stderr -c log_min_duration_statement=0
