+++ package.json
@@ -7,7 +7,7 @@
     "test": "test"
   },
   "scripts": {
-    "test": "DATABASE_URL=postgresql://shamiri:shamiri@localhost:5432/rafi_db yarn build:ts && tsc -p test/tsconfig.json && tap --jobs=1 --ts \"test/**/*.test.ts\"",
+    "test": "DATABASE_URL=postgresql://shamiri:shamiri@localhost:5433/rafi_db yarn build:ts && tsc -p test/tsconfig.json && tap --jobs=1 --ts \"test/**/*.test.ts\"",
     "start": "yarn build:ts && fastify start -l info dist/src/app.js",
     "prod:start": "fastify start -l info dist/src/app.js",
     "build:ts": "tsc",
