+++ src/routes/auth/create-user/index.ts
@@ -1,22 +1,11 @@
 import { FastifyPluginAsync } from "fastify";
 import { Type, Static } from "@sinclair/typebox";
 import { eq, or } from "drizzle-orm";
-import {
-  userAchievement,
-  human,
-  user,
-  userRewardHub,
-  rewardHubRecord,
-  userGoal,
-  userService,
-  subscription,
-  referralCodes,
-} from "../../../database/schema";
-import subscriptionTypes from "../../../../static/subscription_types.json";
+import { human } from "../../../database/schema";
 import { sendVerificationCode } from "../../../lib/auth";
 import { addUserToStream } from "../../../lib/stream";
-import { addDays } from "date-fns";
 import { UserResponse } from "../../../lib/schemas";
+import { createUser } from "../../../services/userService";
 
 const CreateUserBody = Type.Object({
   email: Type.String(), // FIXME: tighten this to use 'email format'
@@ -88,213 +77,7 @@ const createUserRoute: FastifyPluginAsync = async (fastify): Promise<void> => {
         );
       }
 
-      const isMkuUser = email.endsWith("@mylife.mku.ac.ke");
-      const isMoringaUser = email.endsWith("@moringaschool.com");
-      const isZerakiUser = email.endsWith("@zeraki.app");
-      const isAHNUser = email.endsWith("@africahealthcarenetwork.com")
-      const isBelvaUser = email.endsWith("@belvadigital.com")
-
-      // TODO: create a better programmatic way of checking this
-      const MKU_CLIENT_ID = 20;
-      const MORINGA_CLIENT_ID = 18;
-      const ZERAKI_CLIENT_ID = 2;
-      const AHN_CLIENT_ID = 24;
-      const BELVA_CLIENT_ID = 23;
-
-      const SYMON_ID = 10;
-      const HELLEN_ID = 205;
-
-      const userResult = await fastify.db.transaction(async (trx) => {
-        const now = new Date();
-        try {
-          const insertedHumanResult = await trx
-            .insert(human)
-            .values({
-              mobile: phoneNumber,
-              email,
-              role: "user",
-            })
-            .returning();
-
-          // TODO: we need to remove this/make it cleaner
-          const [year, month, date] = request.body.birth_date.split("/");
-          const dateOfBirth = `${year}-${month}-${date}`;
-          let insertedUserResult = await trx
-            .insert(user)
-            .values({
-              id: insertedHumanResult[0].id,
-              dateOfBirth,
-              educationalLevel: request.body.education_level,
-              pinH: Buffer.from(
-                "$2b$12$geh5R2I.08scNPuug5JnRuf/XXS1JsUKKwXAmz9FWb2BrnA/4Pj5G",
-              ),
-              clientId: null,
-              avatarId: 1,
-              gender2: request.body.gender,
-              profession: request.body.profession,
-              registeredOn: now,
-            })
-            .returning();
-
-          const userRewardHubRecordResult = await trx
-            .insert(userRewardHub)
-            .values({
-              userId: insertedHumanResult[0].id,
-              level: 1,
-              gemsHave: 5,
-            })
-            .returning();
-
-          // DEPRECATED: create reward hub record
-          await trx
-            .insert(rewardHubRecord)
-            .values({
-              userRewardHubId: userRewardHubRecordResult[0].id,
-              level: 1,
-              gemsHave: 5,
-              timestamp: now,
-              streak: 0,
-            })
-            .returning();
-
-          // DEPRECATED: create UserGoal record not to be confused with the Goals table
-          await trx.insert(userGoal).values({
-            userRewardHubId: userRewardHubRecordResult[0].id,
-            timestamp: now,
-          });
-
-          // create user achievement record
-          await trx.insert(userAchievement).values({
-            userRewardHubId: userRewardHubRecordResult[0].id,
-            userId: insertedHumanResult[0].id,
-          });
-
-          let userServiceRecord: typeof userService.$inferInsert = {
-            userId: insertedHumanResult[0].id,
-          };
-
-          if (request.body.referral_code) {
-            fastify.log.info(
-              `Referral code provided: ${request.body.referral_code}`,
-            );
-            const referralRecord = await trx.query.referralCodes.findFirst({
-              where: eq(
-                referralCodes.referralCode,
-                request.body.referral_code.trim().toUpperCase(),
-              ),
-            });
-
-            if (referralRecord) {
-              insertedUserResult = await trx
-                .update(user)
-                .set({
-                  clientId: referralRecord.clientId,
-                  referralRecordId: referralRecord.id,
-                })
-                .where(eq(user.id, insertedHumanResult[0].id))
-                .returning();
-
-              userServiceRecord.assignedTherapistId =
-                Math.random() > 0.5 ? HELLEN_ID : SYMON_ID;
-              // TODO: later on we need to toggle if we are using client's own therapists or our own
-            } else {
-              fastify.log.warn(
-                `No referral record found for: ${request.body.referral_code}`,
-              );
-            }
-          } else if (isMkuUser) {
-            fastify.log.info("MKU user identified");
-            insertedUserResult = await trx
-              .update(user)
-              .set({
-                clientId: MKU_CLIENT_ID,
-              })
-              .where(eq(user.id, insertedHumanResult[0].id))
-              .returning();
-
-            userServiceRecord.assignedTherapistId =
-              Math.random() > 0.5 ? HELLEN_ID : SYMON_ID;
-          } else if (isMoringaUser) {
-            fastify.log.info("MORINGA USER INDENTIFIED");
-            if (request.body.referral_code) {
-              fastify.log.info("REFERRAL RECORD ALSO DETECTED");
-            }
-            insertedUserResult = await trx
-              .update(user)
-              .set({
-                clientId: MORINGA_CLIENT_ID,
-              })
-              .where(eq(user.id, insertedHumanResult[0].id))
-              .returning();
-
-            userServiceRecord.assignedTherapistId =
-              Math.random() > 0.5 ? HELLEN_ID : SYMON_ID;
-          } else if (isZerakiUser) {
-            fastify.log.info("ZERAKI USER IDENTIFIED");
-            insertedUserResult = await trx
-              .update(user)
-              .set({
-                clientId: ZERAKI_CLIENT_ID,
-              })
-              .where(eq(user.id, insertedHumanResult[0].id))
-              .returning();
-
-            userServiceRecord.assignedTherapistId =
-              Math.random() > 0.5 ? HELLEN_ID : SYMON_ID;
-          } else if (isBelvaUser) {
-            fastify.log.info("BELVA USER IDENTIFIED")
-            insertedUserResult = await trx
-              .update(user)
-              .set({
-                clientId: BELVA_CLIENT_ID,
-              })
-              .where(eq(user.id, insertedHumanResult[0].id))
-              .returning();
-
-            userServiceRecord.assignedTherapistId =
-              Math.random() > 0.5 ? HELLEN_ID : SYMON_ID;
-          } else if (isAHNUser) {
-            fastify.log.info("AHN USER IDENTIFIED")
-            insertedUserResult = await trx
-              .update(user)
-              .set({
-                clientId: AHN_CLIENT_ID,
-              })
-              .where(eq(user.id, insertedHumanResult[0].id))
-              .returning();
-
-            userServiceRecord.assignedTherapistId =
-              Math.random() > 0.5 ? HELLEN_ID : SYMON_ID;
-          }
-
-          await trx.insert(userService).values(userServiceRecord);
-
-          if (!insertedUserResult[0].clientId) {
-            const { validity, credit, subscriptionType } =
-              subscriptionTypes.subscriptionOrder.individualIntroFreemium;
-
-            const expireTime = addDays(now, validity);
-            await trx.insert(subscription).values({
-              userId: insertedHumanResult[0].id,
-              type: subscriptionType,
-              timestamp: now,
-              expireTime,
-              ref: phoneNumber,
-              totalCredit: credit,
-              remCredit: credit,
-            });
-          }
-
-          return {
-            ...insertedUserResult[0],
-            name: insertedHumanResult[0].name,
-          };
-        } catch (error) {
-          fastify.log.error(error);
-          await trx.rollback();
-          throw error;
-        }
-      });
+      const userResult = await createUser(fastify.db, request.body);
 
       if (userResult) {
         try {
